requirements.txt
aiogram==3.0.0b7
gspread
oauth2client==4.1.3
pandas==2.1.4
matplotlib==3.8.2
openpyxl==3.1.2
XlsxWriter==3.1.9
python-dotenv==1.0.0
Pillow==10.1.0